requirements.txt
Django==5.1.4
drf-spectacular==0.28.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
psycopg2==2.9.10
django-redis==5.4.0